+++ backend/Models/Product.cs
@@ -10,7 +10,7 @@ namespace backend.Models;
 public class Product : BaseModel
 {
   public string? Name { get; set; }
-  public float? Price { get; set; } 
+  public decimal? Price { get; set; } 
   public string? Description { get; set; } 
   public string? Image { get; set; }
   public int CategoryId { get; set; }
